fix: split image-url.ts to prevent fs require in client bundle (#212)
Move fs-dependent functions (loadImageIndex, getManifestHeroImage,
getHeroImage) to image-url.server.ts with 'server-only' guard.
Client-safe image-url.ts retains only resolveImageUrl and
resolveImageUrlClient — no Node.js module references.

Server components (page.tsx, hot/page.tsx) now import from
image-url.server; client components continue importing from
image-url.ts unchanged.

Addresses issue #212
